Simplify favicons, touch icons, etc. and reduce number of images!
A huge reduction of necessary number of images! Support will be a bit less (WIndows 8?!), and there’s only one apple-touch-icon left, that will be scaled down on older, non-retine iOS devices; same for the favicon for all modern browsers: just one 192x192 pixel PNG that will be scaled down. So for pixel perfect control peep, smaller versions of the PNG favicon need to be created and link added in HTML head… but probably for 96% of the use cases this will be (more than) good enough ;)